+++ app/src/main/java/com/gallery/sample/enums/ScanType.kt
@@ -1,0 +1,9 @@
+package com.gallery.sample.enums
+
+enum class ScanType {
+    FILE,
+    AUDIO,
+    IMAGE,
+    VIDEO,
+    MIX
+}