+++ scan/src/main/java/com/gallery/scan/Scan.kt
@@ -5,9 +5,9 @@ import android.os.Bundle
 interface Scan {
 
     /** 扫描多个数据 */
-    fun scanMultiple(bundle: Bundle)
+    fun scanMultiple(args: Bundle)
 
     /** 扫描单个数据 */
-    fun scanSingle(bundle: Bundle)
+    fun scanSingle(args: Bundle)
 
 }