+++ scan/src/main/java/com/gallery/scan/args/audio/AudioColumns.kt
@@ -1,7 +1,6 @@
 package com.gallery.scan.args.audio
 
 import android.net.Uri
-import android.provider.BaseColumns
 import android.provider.MediaStore
 
 internal object AudioColumns {
@@ -9,7 +8,7 @@ internal object AudioColumns {
     internal val uri: Uri = MediaStore.Audio.Media.EXTERNAL_CONTENT_URI
 
     internal val columns: Array<String> = arrayOf(
-            BaseColumns._ID,
+            MediaStore.Audio.Media._ID,
             MediaStore.Audio.Media.SIZE,
             MediaStore.Audio.Media.DISPLAY_NAME,
             MediaStore.Audio.Media.TITLE,