+++ scan/src/main/java/com/gallery/scan/args/file/ScanFileArgs.kt
@@ -11,27 +11,25 @@ import kotlinx.android.parcel.Parcelize
 
 /**
  * 文件扫描
- * 默认只扫描图片
  * 根据[MediaStore.Files.FileColumns.DATE_MODIFIED]字段[Sort.DESC]排序
  */
 @Parcelize
 class ScanFileArgs(
-        private val scanTypeArray: Array<String> = arrayOf(MediaStore.Files.FileColumns.MEDIA_TYPE_IMAGE.toString()),
+        private val scanTypeArray: Array<String>?,
         private val scanSortField: String = MediaStore.Files.FileColumns.DATE_MODIFIED,
         private val scanSort: String = Sort.DESC,
 ) : CursorLoaderArgs(FileColumns.uri, FileColumns.columns, "$scanSortField $scanSort") {
 
     override fun createSelection(args: Bundle): String? {
         val parent: Long = args.getLong(MediaStore.Files.FileColumns.PARENT)
+        scanTypeArray ?: return null
         return when (args.getSerializable(MediaStore.Files.FileColumns.MIME_TYPE) as Result) {
             Result.SINGLE -> resultSelection(args.getLong(MediaStore.Files.FileColumns._ID), scanTypeArray)
             Result.MULTIPLE -> if (parent == SCAN_ALL) scanTypeSelection(scanTypeArray) else parentSelection(parent, scanTypeArray)
         }
     }
 
-    override fun createSelectionArgs(args: Bundle): Array<String>? {
-        return scanTypeArray
-    }
+    override fun createSelectionArgs(args: Bundle): Array<String>? = scanTypeArray
 
     companion object {
 
@@ -43,6 +41,9 @@ class ScanFileArgs(
 
         private fun scanTypeSelection(scanTypeArray: Array<String>): String {
             val defaultSelection = StringBuilder(" ${MediaStore.Files.FileColumns.SIZE} > 0 AND ")
+            if (scanTypeArray.isNullOrEmpty()) {
+                return defaultSelection.toString().removeSuffix("AND ")
+            }
             scanTypeArray.forEach { _ -> defaultSelection.append(APPEND) }
             return defaultSelection.toString().removeSuffix("OR")
         }